Dynamic pickup points. Physical hands. Abstraction. Collider deregistration.
Added the option to have hands be physical. Needs to be switched to a
separate object instead of added to the controller. Physical hands could
be used for buttons and such. Abstracted out VRInteractable a bit. This
way you can have hinged joints that inherit from vrinteractable but not
interactable item. Items now deregister themselves on deletion. Fixed
some other bugs.